LEADS SCORING  - case study
A Brief Overview of The Problem Statement: -
X Education, an Education company, sells online courses to 
Professionals. They have abundance of potential leads who visit their 
website enquiring about the course. But the problem lies in the fact that 
around 70% of those potential leads don’t pan out (i.e., Don’t enrol in 
their course). 
The company has hired our team of Data Analysts to pinpoint the traits 
that a lead shows to ascertain, with a respectable accuracy, if the lead 
will turn into an actual customer or not. We are required to build a 
model that gives these leads a score between 0-100. This will help their 
sales team to focus on only the clients which have a high leads score.